Removed closing PHP tag in class file
Remove closing PHP tag to follow best practices.

"If a file is pure PHP code, it is preferable to omit the PHP closing
tag at the end of the file. This prevents accidental whitespace or new
lines being added after the PHP closing tag, which may cause unwanted
effects because PHP will start output buffering when there is no
intention from the programmer to send any output at that point in the
script."

References:
- http://php.net/manual/en/language.basic-syntax.phptags.php
- https://www.drupal.org/coding-standards#phptags
-

http://framework.zend.com/manual/current/en/ref/coding.standard.html#general